Refactor impl (#36)
* Refactor impl

* Refactor impl 2
