+++ NukeBuildHelpers/BaseNukeBuildHelpers.Targets.User.Unlisted.cs
@@ -1,5 +1,6 @@
 ﻿using Nuke.Common;
 using Nuke.Common.Tooling;
+using NukeBuildHelpers.Common;
 using Serilog;
 
 namespace NukeBuildHelpers;
@@ -11,6 +12,8 @@ partial class BaseNukeBuildHelpers
         .Description("Delete all origin tags, with --args \"{appid}\"")
         .Executes(() =>
         {
+            CheckEnvironementBranches();
+
             List<string> tagsToDelete = [];
             if (string.IsNullOrEmpty(Args))
             {
@@ -23,8 +26,8 @@ partial class BaseNukeBuildHelpers
             }
             else
             {
-                GetOrFail(() => SplitArgs, out var splitArgs);
-                GetOrFail(() => GetAppConfig(), out var appConfig);
+                ValueHelpers.GetOrFail(() => SplitArgs, out var splitArgs);
+                ValueHelpers.GetOrFail(() => AppEntryHelpers.GetAppConfig(), out var appConfig);
 
                 IReadOnlyCollection<Output>? lsRemote = null;
 
@@ -32,8 +35,8 @@ partial class BaseNukeBuildHelpers
                 {
                     string appId = key;
 
-                    GetOrFail(appId, appConfig.AppEntryConfigs, out appId, out var appEntry);
-                    GetOrFail(() => GetAllVersions(appId, appConfig.AppEntryConfigs, ref lsRemote), out var allVersions);
+                    ValueHelpers.GetOrFail(appId, appConfig.AppEntryConfigs, out appId, out var appEntry);
+                    ValueHelpers.GetOrFail(() => AppEntryHelpers.GetAllVersions(this, appId, appConfig.AppEntryConfigs, ref lsRemote), out var allVersions);
 
                     if (appEntry.Entry.MainRelease)
                     {
